// src/generated/transferSyntaxes.ts
export type PackedTransferSyntax = readonly [
    uid: string,
    name: string,
    keyword: string | null,
    dcmtkName: string | null,
    explicitVr: boolean,
    byteOrder: string,
    pixelDataEncoding: string,
    pixelDataCompression: string,
    fragmentable: boolean,
    streamCompression: string,
    status: string,
    mimeType: string,
    filenameExtension: string,
];